Menu reorder, added conditional reactions
Biggest change i made was changing the order of instantiation for the different menus since reactions and bonus actions can be conditional. To this effect i put the reactions and bonus actions first. I added the reaction button and reaction menu for keeping track of reactions taken. I used a menu instead of simply graying out the button since there are certain reactions you can take during your turn. Also added the use readied action as well as making it conditional if ready action was used. I added a reset turn button to reset the turn and the buttons to default states and divided the options menu and reset button into the left and right part of an HBox respectively.